Add files via upload
E-Commerce contract:
E-Commerce company that sells clothing online but also have in-store style and clothing sessions. Customers come in to the store, have sessions/meetings with a personel stylist, then they can go home and order either on mobile app or website for cloth thet want.

The company is trying to decide whether to focus on their efforts on mobile app experience or their website. 

They are providing customer data for us to analyze and conclude.

**data is in csv Format: File- Ecommerce Customers